requirements.txt
Pillow~=8.3.2
tkinterdnd2~=0.3.0